requirements.txt
openai
google-genai
numpy
pandas
matplotlib
